This time send edginess plots without titles.
The raw profile has single-width text for the abscissa times and
part of the ordinate label, athough it was produced in the same
fashion as the other plot.  On the previous version with the title
this behaviour reset to thick lines at normal magnification.  This time
the the thin lines are present at all magnifications.