Make a first stab at replacing the remaining P61 plot.
This is just a first go.  I couldn't find an example of first periodic
signal as shown in the previous plot (lower graphic in P61_f2a).  If
deemed important, I could add the offsets of the red curves, and the
number of spectra used to form them; and the non-linearity figure for
each plot, as in the text I do want to give some quantitative examples
of good and bad and the thresholds used.