Figure 5:  Changed labelling
 + Bigger labels
 + Centred ordinate label spannig both cells
 + Remove upper plot's abscissa numeric labels

The reason the text was so small before was to accommodate the
ordinate text label within on of the cells.  The current plot was
something of a fudge to get the label where I wanted as there are
no offsets for the location of textual labels _parallel_ to each's
respective axis.  This meant that I could quite avoid a gap between
the two graphs.  Perhaps with an hour or two more fudging, it might be
possible, but there are other plots to revise.