Figure 12: Replaced original graphics.
The delay has been because I was unable to reproduce the before image
of L1455.  So I found a different example in OrionA.  The effects are
more subtle but also more typical.

 + Scale both panels to the same limits and display in key
 + Add axis labels and sky co-ordinates.

Things I don't like is the DISPLAY key and the inability to make it
taller and a little wider.  Also the unflatfield image looks brighter
than the cleaned up picture as if there's an offset in intensity.
Presenting the integ adds complications of how the sources are extracted
in each case.  I also had to regenerate the data, so likely there were
other differences, and perhaps I should regenerate the right-panel
integ.