Figure 12: Have another go.
The previous attempt was based upon a file in a noflat/directory, but
that was misleading.  It had been flatfielded using the default
recpars for GBS, not the copy with flatfielding disabled.  The
velocity limits used were unsuitable, making the integ worse.  The
crazy ratios lead to an overall brightening by 50%.

This is a Taurus field instead.  Yes it's at an angle to all the
regular sky co-ordinate systems.  This means the tick marks had to
change to black to make them stand out.  They are yellow in other
plots to contrast with the dark cubehelix0 background.

A bonus is that the key is big again.